requirements.txt
mpi4py
numpy
matplotlib
pycuda
biopython